+++ common/src/main/java/com/github/jelatinone/acquisition/fetch/HTTPMetadataFetcher.java
@@ -54,7 +54,7 @@ public class HTTPMetadataFetcher implements MetadataFetcher {
         Long contentLength = response.headers().firstValue("Content-Length")
             .map((length) -> {
               try {
-                return (long) Long.parseLong(length);
+                return Long.parseLong(length);
               } catch (Exception exception) {
                 return null;
               }
